requirements.txt
Django==2.1.4
Pillow==5.3.0
psycopg2==2.7.6.1
pytz==2018.7
